Use the native Touchable implementation on all native platforms
